sbux_example/requirements.txt
azure-functions
jsons
slack_webhook
Slack
